pbird_backend/requirements.txt
asgiref==3.5.0
astroid==2.11.4
bidict==0.21.4
black==22.3.0
certifi==2021.10.8
cffi==1.15.0
cfgv==3.3.1
charset-normalizer==2.0.11
click==8.1.3
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.1
dill==0.3.4
distlib==0.3.4
Django==3.2.5
django-ckeditor==6.2.0
django-cors-headers==3.7.0
django-environ==0.8.1
django-filter==21.1
django-js-asset==2.0.0
django-pdfkit==0.3.1
django-phonenumber-field==6.0.0
django-rest-knox==4.1.0
djangorestframework==3.12.4
djongo==1.3.6
dnspython==2.2.0
drf-yasg==1.20.0
filelock==3.4.2
fpdf==1.7.2
fpdf2==2.5.0
identify==2.5.0
idna==3.3
inflection==0.5.1
isort==5.10.1
itypes==1.2.0
Jinja2==3.0.3
lazy-object-proxy==1.7.1
MarkupSafe==2.0.1
mccabe==0.7.0
mongoengine==0.23.1
mypy-extensions==0.4.3
nodeenv==1.6.0
packaging==21.3
pathspec==0.9.0
pdfkit==0.6.0
phonenumbers==8.12.22
Pillow==8.3.1
platformdirs==2.4.1
pre-commit==2.18.1
pre-commit-hooks==4.2.0
pycparser==2.21
PyJWT==1.7.1
pylint==2.13.8
pymongo==3.12.3
pyparsing==3.0.7
python-barcode==0.13.1
python-decouple==3.6
python-engineio==4.3.1
python-http-client==3.3.7
python-snappy==0.6.0
python-socketio==5.5.2
pytz==2021.3
PyYAML==6.0
qrcode==6.1
requests==2.27.1
ruamel.yaml==0.17.20
ruamel.yaml.clib==0.2.6
sendgrid==6.9.7
six==1.16.0
sqlparse==0.2.4
starkbank-ecdsa==2.0.3
toml==0.10.2
tomli==2.0.1
twilio==6.59.0
typing_extensions==4.2.0
tzdata==2021.5
uritemplate==4.1.1
urllib3==1.26.8
virtualenv==20.13.1
virtualenvwrapper-win==1.2.7
websocket-client==1.3.1
websockets==10.2
wrapt==1.14.1
